github: publish workspace folder
